Accept fractional local model VRAM estimates
Allow fractional model size estimates while rejecting values that cannot form a positive finite budget. Change the `LocalModelConfig` field `vram_gb` and its accessor from `Option<u32>` to `Option<f64>`, then validate each configured estimate in `validate.rs`.
- Store the estimate as `f64` so validation and co-residency sums use one numeric type.
- Validation rejects zero, negative, NaN, and infinite estimates with `vram_gb must be finite and greater than zero`. Tests pin fractional input, integer compatibility, and invalid-value rejection.

+++ crates/promptforge-gateway-config/src/config/accessors.rs
@@ -1369,11 +1369,11 @@ impl LocalModelConfig {
     /// # vram_gb = 14
     /// # "#;
     /// let config = Config::from_toml_str(toml)?;
-    /// assert_eq!(config.local_models()[0].vram_gb(), Some(14));
+    /// assert_eq!(config.local_models()[0].vram_gb(), Some(14.0));
     /// # Ok::<(), promptforge_gateway_config::ConfigError>(())
     /// ```
     #[must_use]
-    pub fn vram_gb(&self) -> Option<u32> {
+    pub fn vram_gb(&self) -> Option<f64> {
         self.vram_gb
     }
